ENH: add a timing-not-changed counter
each time a benchmark is rerun in `min` mode and the timing is not
lower than the already recorded one a counter is increased.
If the counter is above `unchanged_rerun_limit` the benchmark is not
rerun again. This avoids rerunning stable benchmarks.
If the timing is faster than the recorded one the counter is reset to
zero.